moved HealthCheckListener logic into CarinaListener
